+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/signaling/message/MessageConverter.kt
@@ -4,6 +4,7 @@ import com.google.gson.Gson
 import jp.shiguredo.sora.sdk.channel.option.SoraChannelRole
 import jp.shiguredo.sora.sdk.channel.option.SoraMediaOption
 import jp.shiguredo.sora.sdk.util.SoraLogger
+import java.util.*
 
 class MessageConverter {
 
@@ -18,17 +19,19 @@ class MessageConverter {
                                 channelId:               String?,
                                 mediaOption:             SoraMediaOption,
                                 metadata:                Any?,
-                                sdp:                     String,
+                                sdp:                     String?          = null,
+                                sdpError:                String?          = null,
                                 clientId:                String?          = null,
                                 signalingNotifyMetadata: Any?             = null
         ): String {
 
             val msg = ConnectMessage(
-                    role                    = role.toString().toLowerCase(),
+                    role                    = role.toString().toLowerCase(Locale.getDefault()),
                     channelId               = channelId,
                     metadata                = metadata,
                     multistream             = mediaOption.multistreamIsRequired,
                     sdp                     = sdp,
+                    sdp_error               = sdpError,
                     clientId                = clientId,
                     signalingNotifyMetadata = signalingNotifyMetadata
             )